+++ tests/testthat/test-fill_gaps.R
@@ -24,16 +24,6 @@ test_that("fill_down_until limits fills to max timeslots", {
   expect_true(is.na(filled$temperature[5]))
 })
 
-test_that("complete_timeseries_datetime reconstructs missing timestamps", {
-  dtf_gaps <- timefully::dtf[c(1, 2, 3, 8, 9, 10), ]
-  filled <- complete_timeseries_datetime(dtf_gaps)
-
-  expect_equal(
-    nrow(filled),
-    96 # Full day
-  )
-})
-
 test_that("fill_na works", {
   past_data <- data.frame(
     datetime = as.POSIXct("2024-01-01 00:00:00", tz = "UTC") + 0:3 * 3600,